messi

+++ degrees.cpp
@@ -6,19 +6,19 @@ Degrees::Degrees(){
     this->degrees = 0.0;
 }
 
-Degrees::Degrees(double degrees){
+Degrees::Degrees(float degrees){
     this->degrees = fabs(fmod(degrees, 360.0));
 }
 
 
 
-void Degrees::operator= (const double & degrees){
+void Degrees::operator= (const float & degrees){
     this->degrees = fabs(fmod(degrees, 360.0));
 }
 
 
 
-void Degrees::operator-= (const double & degrees){
+void Degrees::operator-= (const float & degrees){
     this->degrees -= degrees;
     this->degrees = fabs(fmod(degrees, 360.0));
 }
@@ -30,7 +30,7 @@ void Degrees::operator-= ( Degrees  degrees){
 
 
 
-void Degrees::operator+= (const double & degrees){
+void Degrees::operator+= (const float & degrees){
     this->degrees += degrees;
     this->degrees = fabs(fmod(degrees, 360.0));
 }
@@ -42,8 +42,8 @@ void Degrees::operator+= ( Degrees  degrees){
 
 
 
-bool Degrees::operator== (const double & degrees){
-   double temp =   fabs(fmod(degrees, 360.0));
+bool Degrees::operator== (const float & degrees){
+   float temp =   fabs(fmod(degrees, 360.0));
     if(this->degrees == temp)
         return true;
     else
@@ -52,8 +52,8 @@ bool Degrees::operator== (const double & degrees){
 
 
 
-bool Degrees::operator< (const double & degrees){
-    double temp =   fabs(fmod(degrees, 360.0));
+bool Degrees::operator< (const float & degrees){
+    float temp =   fabs(fmod(degrees, 360.0));
     if(this->degrees < temp)
         return true;
     else
@@ -61,8 +61,8 @@ bool Degrees::operator< (const double & degrees){
 }
 
 
-bool Degrees::operator> (const double & degrees){
-    double temp =   fabs(fmod(degrees, 360.0));
+bool Degrees::operator> (const float & degrees){
+    float temp =   fabs(fmod(degrees, 360.0));
     if(this->degrees > temp)
         return true;
     else
@@ -70,8 +70,8 @@ bool Degrees::operator> (const double & degrees){
 }
 
 
-bool Degrees::operator>= (const double & degrees){
-   double temp =   fabs(fmod(degrees, 360.0));
+bool Degrees::operator>= (const float & degrees){
+   float temp =   fabs(fmod(degrees, 360.0));
     if(this->degrees >= temp)
         return true;
     else
@@ -80,15 +80,15 @@ bool Degrees::operator>= (const double & degrees){
 
 
 
-bool Degrees::operator<= (const double & degrees){
-    double temp =   fabs(fmod(degrees, 360.0));
+bool Degrees::operator<= (const float & degrees){
+    float temp =   fabs(fmod(degrees, 360.0));
     if(this->degrees <= temp)
         return true;
     else
         return false;
 }
 
-Degrees Degrees::operator+ (const double & degrees){
+Degrees Degrees::operator+ (const float & degrees){
     Degrees temp = *this;
     temp += degrees;
     temp.degrees = fabs(fmod(temp.degrees, 360.0));
@@ -96,7 +96,7 @@ Degrees Degrees::operator+ (const double & degrees){
 }
 
 
-Degrees & Degrees::operator- (const double & degrees){
+Degrees & Degrees::operator- (const float & degrees){
     this-> degrees -= degrees;
     this->degrees = fabs(fmod(degrees, 360.0));
     return *this;
@@ -108,7 +108,7 @@ Degrees Degrees::operator+ ( Degrees degrees){
     return *this ;
 }
 
-double Degrees::getDegrees(){
+float Degrees::getDegrees(){
     return this->degrees;
 }
 
@@ -120,7 +120,7 @@ Degrees & Degrees::operator- ( Degrees  degrees){
 
 
 
-double Degrees::getRadians(){
+float Degrees::getRadians(){
     return this->degrees * PI/180.0;
 }
