+++ point.cpp
@@ -2,7 +2,7 @@
 namespace SSJServer {
 
 
-    Point::Point(double x, double y){
+    Point::Point(float x, float y){
         this->x = x;
         this->y = y;
     }
